restructured project directory
