+++ package.json
@@ -9,19 +9,21 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/metro-runtime": "~4.0.0",
     "@react-navigation/bottom-tabs": "^7.0.3",
     "@react-navigation/elements": "^2.0.2",
     "@react-navigation/native": "^7.0.2",
     "@react-navigation/native-stack": "^7.0.2",
     "expo": "~52.0.6",
+    "expo-linking": "~7.0.2",
     "expo-status-bar": "~2.0.0",
     "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-native": "0.76.1",
+    "react-native-paper": "^5.12.5",
     "react-native-safe-area-context": "^4.12.0",
     "react-native-screens": "~4.0.0",
-    "react-dom": "18.3.1",
-    "react-native-web": "~0.19.13",
-    "@expo/metro-runtime": "~4.0.0"
+    "react-native-web": "~0.19.13"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
