+++ src/pages/Scanner.js
